refactor mappers in data
